feat:(video-web)订单相关接口&商品信息相关接口

+++ video-web/src/main/java/com/example/videoweb/domain/entity/Product.java
@@ -44,7 +44,7 @@ public class Product {
     private String thumb;
 
     @Schema(description = "商品评分")
-    private Long rate;
+    private Float rate;
 
     @Schema(description = "商品价格")
     private Long price;
